Changed price to float

+++ Assets/Scripts/Inventory/ItemData.cs
@@ -7,6 +7,6 @@ public class ItemData : ScriptableObject
     public string Name;
     public string Description;
     public Sprite Icon;
-    public int Price;
+    public float Price;
     public bool IsStackable;
 }
